Former-commit-id: bc5c895f70a97a9c6037768489a668b8bda6f28a

+++ rs-dam/indexer/indexer-domain/src/main/java/fr/cnes/regards/modules/indexer/domain/DataFile.java
@@ -20,9 +20,9 @@ package fr.cnes.regards.modules.indexer.domain;
 
 import java.net.URI;
 
+import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
-import javax.validation.constraints.NotBlank;
 import org.springframework.util.MimeType;
 
 import fr.cnes.regards.framework.oais.urn.DataType;
@@ -62,12 +62,12 @@ public class DataFile {
     /**
      * Required width if image file
      */
-    private Integer imageWidth;
+    private Double imageWidth;
 
     /**
      * Required height if image file
      */
-    private Integer imageHeight;
+    private Double imageHeight;
 
     /**
      * Required field to know if the file is online ? (if not, it is NEARLINE)
@@ -138,19 +138,19 @@ public class DataFile {
         this.mimeType = mimeType;
     }
 
-    public Integer getImageWidth() {
+    public Double getImageWidth() {
         return imageWidth;
     }
 
-    public void setImageWidth(Integer imageWidth) {
+    public void setImageWidth(Double imageWidth) {
         this.imageWidth = imageWidth;
     }
 
-    public Integer getImageHeight() {
+    public Double getImageHeight() {
         return imageHeight;
     }
 
-    public void setImageHeight(Integer imageHeight) {
+    public void setImageHeight(Double imageHeight) {
         this.imageHeight = imageHeight;
     }
